fusion drag context and chat context... chat now stay open if already opened before switching page
